Refactor encoding packages to be nested under common package
Updates the encoding packages, time, httpbinding, json and xml to all be
nested under a common package, encoding
