+++ encoding/json/encoder_test.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"testing"
 
-	"github.com/awslabs/smithy-go/json"
+	"github.com/awslabs/smithy-go/encoding/json"
 )
 
 func TestEncoder(t *testing.T) {
