+++ encoding/xml/array_test.go
