+++ encoding/xml/constants.go
