+++ encoding/xml/escape.go
