+++ encoding/xml/map_test.go
